backend/pyproject.toml
[project]
name = "comradarr"
version = "0.0.0"
description = "A media library completion service that integrates with *arr applications."
license = "AGPL-3.0-or-later"
authors = [
  { name = "engels74", email = "141435164+engels74@users.noreply.github.com" },
]
requires-python = ">=3.14,<3.15"
dependencies = [
  # Phase 1 runtime dependencies (plan §1 Step 10). httpx[http2] is intentionally
  # NOT listed here — Phase 7 owns its addition together with the first real
  # outbound client (Q6 / C6 final).
  "litestar[standard]>=2.19,<3",
  "granian[uvloop]>=2.7,<3",
  "msgspec>=0.20,<0.21",
  "sqlalchemy[asyncio]>=2.0.46,<2.1",
  "asyncpg>=0.30",
  "alembic>=1.18",
  "structlog>=25.5",
  # advanced-alchemy carries no upper pin per plan — track upstream stability.
  "advanced-alchemy",
  # Phase 3 (§5.3) crypto + audit primitives. The cryptography lower bound is
  # pinned at >=46.0.7 because the 45.x line and 46.0–46.0.6 carry three
  # fixed CVEs (CVE-2026-26007 fix in 46.0.5, CVE-2026-34073 fix in 46.0.6,
  # CVE-2026-39892 fix in 46.0.7); cp314 wheels are available across the
  # whole 46.x series. The upper bound <47 caps the next major.
  "cryptography>=46.0.7,<47",
  "argon2-cffi>=23,<26",
]

[project.scripts]
# `uv run comradarr` (post-install) launches the canonical Granian process.
# `uv run migrate` (Phase 2 onward) applies Alembic migrations to head; the
# CLI owns its own asyncio loop and is exempt from RULE-ASYNC-002.
comradarr = "comradarr.__main__:main"
migrate = "comradarr.cli.migrate:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "ruff",
  "basedpyright",
  "pytest",
  "pytest-asyncio",
  "pytest-xdist",
  "pip-audit",
]

[tool.uv.build-backend]
module-name = "comradarr"

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
  "E", "W", "F", "I", "UP", "B", "C4", "SIM", "RET", "TID", "TC",
  "FA", "ASYNC", "N", "PTH", "S",
]

[tool.ruff.lint.per-file-ignores]
# Tests use `assert` (S101) freely.
"tests/**" = ["S101"]
# Role-permission matrix tests parametrize over Base.metadata.tables.keys()
# (a trusted internal Python tuple of identifiers, not external input) and
# build per-table privilege-check SQL via f-strings. S608 is a false positive
# in this single file — the equivalent of `format` here is a quoted identifier
# from a curated allowlist.
"tests/db/test_role_permissions.py" = ["S101", "S608"]
# Partial-index introspection assembles a fully static pg_indexes query via
# multi-line ``+`` concatenation (the only dynamic pieces are SQLAlchemy
# bound parameters ``:tablename`` / ``:indexname``). S608 fires on the
# multi-line concat shape but no identifier interpolation occurs.
"tests/db/test_partial_indexes.py" = ["S101", "S608"]
# Alembic env.py is import-by-side-effect; relax secret/credential heuristics.
"migrations/env.py" = ["S105", "S106"]
# PRD §21 fixes the exception class names (AuthenticationInvalidCredentials,
# ConnectorUnavailable, InternalUnexpected, …); the *-Error suffix N818
# convention does not apply to this canonical hierarchy.
"src/comradarr/errors/**" = ["N818"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true

[tool.basedpyright]
typeCheckingMode = "recommended"
pythonVersion = "3.14"
# Phase 1 (F-4 burndown) flips `tests/` into the typecheck surface alongside `src/`.
include = ["src", "tests"]
# tests/typecheck/ holds intentional Secret-misuse fixtures (Phase 3 §5.3.6)
# consumed by tests/test_secret_typecheck.py via subprocess. They MUST trip
# basedpyright as part of the gate, so the suite-wide pass excludes them.
exclude = ["tests/typecheck"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
  "integration: requires a live Postgres + the per-worker schema fixture (RECIPE-PYTEST-DB)",
  "e2e: launches a real Granian subprocess and binds the hardcoded port (single-worker pass)",
]
